Replace SubscriberManager with file reading in API route
In the test route of our API v1, the SubscriberManager was removed and replaced with file system reading in order to fetch the content of the newly added 'uk.ftl' in bot locales. Therefore, the response of the GET method in this route now returns the content of the 'uk.ftl' file instead of subscriber data.